refactor(state): store → state 레이어 구조 정리
- 디렉토리명 변경: renderer/store/ → renderer/state/
- 파일명 단수형 통일: hooks→hook, queryKeys→queryKey, queryOptions→queryOption
- Valtio 코드 통합: state.ts + actions.ts → store.ts
- 네이밍 변경: treeViewState → treeViewStore
- 문서 업데이트: state.md 파일 구조 및 생성 기준 명시
