+++ apps/desktop/src/renderer/state/app/queryOption.ts
@@ -1,5 +1,5 @@
 import { isInitialized } from '@/renderer/repo/app'
-import { QUERY_KEYS } from './const.js'
+import { QUERY_KEYS } from './queryKey'
 
 export const getInitializationQueryOptions = () => ({
   queryKey: QUERY_KEYS.initialization(),
