--- apps/desktop/src/renderer/store/treeView/index.ts
@@ -1,10 +1,0 @@
-export {
-  useTopNodeId,
-  useFocusedNodeId,
-  useTreeViewNodes,
-  useAddNewNodeAfter,
-  useRemoveNode,
-} from './hooks'
-
-export { setTreeViewTopNodeId, setTreeViewFocusedNodeId } from './actions'
-export { treeViewState } from './state'
